Update listenlive-20100326.js
Radio BE1 doesn't exist anymore, it's Energy Bern now. I added instead the missing station RADIO BERN1 in place of that station. I'm not sure how important this file really is, as there is another one here: https://github.com/liip/Radios/blob/master/www/js/radioapp.db.js which also lists the stations. That file seems actually more up to date. The Problem is, that I'm not able to listen to RADIO BERN1 with this app. Any idea why this happens? (There seems to be a redirect, could that be the problem?)